+++ R/chached_data.R
@@ -3,9 +3,6 @@
 #' @docType data
 #'
 #' @usage data(cadsl)
-#'
-#' @examples
-#' data(cadsl)
 "cadsl"
 
 
@@ -14,9 +11,6 @@
 #' @docType data
 #'
 #' @usage data(cadae)
-#'
-#' @examples
-#' data(cadae)
 "cadae"
 
 #' Cached ADAETTE data generated with \code{seed = 1}
@@ -24,9 +18,6 @@
 #' @docType data
 #'
 #' @usage data(cadaette)
-#'
-#' @examples
-#' data(cadaette)
 "cadaette"
 
 #' Cached ADCM data generated with \code{seed = 1}
@@ -34,9 +25,6 @@
 #' @docType data
 #'
 #' @usage data(cadcm)
-#'
-#' @examples
-#' data(cadcm)
 "cadcm"
 
 #' Cached ADLB data generated with \code{seed = 1}
@@ -44,9 +32,6 @@
 #' @docType data
 #'
 #' @usage data(cadlb)
-#'
-#' @examples
-#' data(cadlb)
 "cadlb"
 
 #' Cached ADVS data generated with \code{seed = 1}
@@ -54,9 +39,6 @@
 #' @docType data
 #'
 #' @usage data(cadqs)
-#'
-#' @examples
-#' data(cadqs)
 "cadqs"
 
 #' Cached ADRS data generated with \code{seed = 1}
@@ -64,9 +46,6 @@
 #' @docType data
 #'
 #' @usage data(cadrs)
-#'
-#' @examples
-#' data(cadrs)
 "cadrs"
 
 #' Cached ADTTE data generated with \code{seed = 1}
@@ -74,9 +53,6 @@
 #' @docType data
 #'
 #' @usage data(cadtte)
-#'
-#' @examples
-#' data(cadtte)
 "cadtte"
 
 #' Cached ADVS data generated with \code{seed = 1}
@@ -84,7 +60,4 @@
 #' @docType data
 #'
 #' @usage data(cadvs)
-#'
-#' @examples
-#' data(cadvs)
 "cadvs"