+++ R/radsl.R
@@ -18,7 +18,8 @@
 #'
 #' @template return_data.frame
 #'
-#' @import dplyr
+#' @importFrom dplyr mutate
+#' @importFrom magrittr %>%
 #' @importFrom stats rchisq reorder rexp rnorm runif setNames
 #' @importFrom tibble tibble
 #'
@@ -27,36 +28,46 @@
 #' @examples
 #' ADSL <- radsl(N = 10, seed = 1)
 #' head(ADSL)
-#'
-#' attr(ADSL, "label") # or also tern::label(ADSL)
-#'
-radsl <- function(N = 400, seed = NULL, cached = FALSE) {
+radsl <- function(N = 400, # nolint
+                  seed = NULL,
+                  cached = FALSE) {
 
-  stopifnot(is.logical(cached))
-  if (cached) return(get_cached_data("cadsl"))
+  stopifnot(is.logical.single(cached))
+  if (cached) {
+    return(get_cached_data("cadsl"))
+  }
 
-  if (!is.null(seed)) set.seed(seed)
+  stopifnot(is.numeric.single(N))
+  stopifnot(is.null(seed) || is.numeric.single(seed))
+
+  if (!is.null(seed)) {
+    set.seed(seed)
+  }
 
   country_site_prob <- c(.5, .121, .077, .077, .075, .052, .046, .025, .014, .003)
 
-  ADSL <- tibble(
+  ADSL <- tibble( # nolint
     STUDYID = rep("AB12345", N),
-    COUNTRY = c("CHN", "USA", "BRA", "PAK", "NGA", "RUS", "JPN", "GBR", "CAN", "CHE") %>%
-      sample_fct(N, prob = country_site_prob),
-    SITEID = 1:20 %>%
-      sample_fct(N,  prob = rep(country_site_prob, times = 2)),
+    COUNTRY = sample_fct(
+      c("CHN", "USA", "BRA", "PAK", "NGA", "RUS", "JPN", "GBR", "CAN", "CHE"),
+      N, prob = country_site_prob
+    ),
+    SITEID = sample_fct(1:20, N,  prob = rep(country_site_prob, times = 2)),
     SUBJID = paste("id", seq_len(N), sep = "-"),
-    AGE = sapply(floor(rnorm(N, mean = 20, sd = 20)), max, 0) + 20,
-    SEX = c("F", "M", "U", "UNDIFFERENTIATED") %>% sample_fct(N, prob = c(.5, .48, .015, .005)),
-    ARMCD = c("ARM A", "ARM B", "ARM C") %>%sample_fct(N),
-    RACE = c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", "AMERICAN INDIAN OR ALASKA NATIVE",
-             "MULTIPLE", "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER", "OTHER", "UNKNOWN") %>%
-      sample_fct(N, prob = c(.55, .23, .16, .05, .004, .003, .002, .002)),
-    STRATA1 = c("A", "B", "C") %>% sample_fct(N),
-    STRATA2 = c("S1", "S2") %>% sample_fct(N),
+    AGE = pmax(floor(rnorm(N, mean = 20, sd = 20)), 0) + 20,
+    SEX = sample_fct(c("F", "M", "U", "UNDIFFERENTIATED"), N, prob = c(.5, .48, .015, .005)),
+    ARMCD = sample_fct(c("ARM A", "ARM B", "ARM C"), N),
+    RACE = sample_fct(
+      c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", "AMERICAN INDIAN OR ALASKA NATIVE",
+        "MULTIPLE", "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER", "OTHER", "UNKNOWN"),
+      N,
+      prob = c(.55, .23, .16, .05, .004, .003, .002, .002)
+    ),
+    STRATA1 = sample_fct(c("A", "B", "C"), N),
+    STRATA2 = sample_fct(c("S1", "S2"), N),
     BMRKR1 = rchisq(N, 6),
-    BMRKR2 = c("LOW", "MEDIUM", "HIGH") %>% sample_fct(N),
-    BEP01FL = c("Y", "N") %>% sample_fct(N)
+    BMRKR2 = sample_fct(c("LOW", "MEDIUM", "HIGH"), N),
+    BEP01FL = sample_fct(c("Y", "N"), N)
   ) %>%
     mutate(ARM = recode(ARMCD, "ARM A" = "A: Drug X", "ARM B" = "B: Placebo", "ARM C" = "C: Combination")) %>%
     mutate(ACTARM = ARM) %>%
@@ -64,13 +75,12 @@ radsl <- function(N = 400, seed = NULL, cached = FALSE) {
     mutate(ITTFL = "Y")
 
   # associate sites with countries
-  ADSL <- ADSL %>%
+  ADSL <- ADSL %>% # nolint
     mutate(SITEID = paste0(COUNTRY, "-", SITEID)) %>%
     mutate(USUBJID = paste(STUDYID, SITEID, SUBJID, sep = "-"))
 
   # apply metadata
-  ADSL <- apply_metadata(ADSL, "metadata/ADSL.yml", seed = seed)
+  ADSL <- apply_metadata(ADSL, "metadata/ADSL.yml", seed = seed) # nolint
 
   ADSL
-
 }