+++ R/radtte.R
@@ -14,6 +14,9 @@
 #'
 #' @template return_data.frame
 #'
+#' @importFrom dplyr filter mutate rowwise
+#' @importFrom magrittr %>%
+#' @importFrom tibble tribble
 #'
 #' @export
 #'
@@ -21,65 +24,77 @@
 #' ADSL <- radsl(seed = 1)
 #' ADTTE <- radtte(ADSL, seed = 2)
 #' head(ADTTE)
-#'
-radtte <- function(ADSL, lookup = NULL, event.descr = NULL, seed = NULL, cached = FALSE) {
+radtte <- function(ADSL, # nolint
+                   lookup = NULL,
+                   event.descr = NULL, # nolint
+                   seed = NULL,
+                   cached = FALSE) {
 
-  stopifnot(is.logical(cached))
-  if (cached) return(get_cached_data("cadtte"))
+  stopifnot(is.logical.single(cached))
+  if (cached) {
+    return(get_cached_data("cadtte"))
+  }
 
-  if (!is.null(seed)) set.seed(seed)
+  stopifnot(is.data.frame(ADSL))
+  stopifnot(is.null(seed) || is.numeric.single(seed))
 
-  lookup_TTE <- if (is.null(lookup)){
+  if (!is.null(seed)) {
+    set.seed(seed)
+  }
+
+  lookup_TTE <- if_null( # nolint
+    lookup,
     tribble(
       ~ARM,  ~PARAMCD, ~PARAM, ~LAMBDA, ~CNSR_P,
-      "ARM A", "OS",  "Overall Survival",          1/80,  0.4,
-      "ARM B", "OS",  "Overall Survival",          1/100, 0.2,
-      "ARM C", "OS",  "Overall Survival",          1/60,  0.42,
-      "ARM A", "PFS", "Progression Free Survival", 1/100, 0.3,
-      "ARM B", "PFS", "Progression Free Survival", 1/150, 0.1,
-      "ARM C", "PFS", "Progression Free Survival", 1/80,  0.32,
-      "ARM A", "EFS", "Event Free Survival",       1/80,  0.2,
-      "ARM B", "EFS", "Event Free Survival",       1/100, 0.08,
-      "ARM C", "EFS", "Event Free Survival",       1/60,  0.23)
-  } else {
-    lookup
-  }
-
-  evntdescr_sel <- if (is.null(event.descr)){
-    c(
-      'Death',
-      'Disease Progression',
-      'Last Tumor Assessment',
-      'Adverse Event',
-      'Last Date Known To Be Alive'
+      "ARM A", "OS",  "Overall Survival",          1 / 80,  0.4,
+      "ARM B", "OS",  "Overall Survival",          1 / 100, 0.2,
+      "ARM C", "OS",  "Overall Survival",          1 / 60,  0.42,
+      "ARM A", "PFS", "Progression Free Survival", 1 / 100, 0.3,
+      "ARM B", "PFS", "Progression Free Survival", 1 / 150, 0.1,
+      "ARM C", "PFS", "Progression Free Survival", 1 / 80,  0.32,
+      "ARM A", "EFS", "Event Free Survival",       1 / 80,  0.2,
+      "ARM B", "EFS", "Event Free Survival",       1 / 100, 0.08,
+      "ARM C", "EFS", "Event Free Survival",       1 / 60,  0.23
     )
-  } else {
-    event.descr
-  }
+  )
 
-  ADTTE <- split(ADSL, ADSL$USUBJID) %>% lapply(FUN = function(pinfo) {
+  evntdescr_sel <- if_null(
+    event.descr,
+    c(
+      "Death",
+      "Disease Progression",
+      "Last Tumor Assessment",
+      "Adverse Event",
+      "Last Date Known To Be Alive"
+    )
+  )
 
-    lookup_TTE %>% filter(ARM == as.character(pinfo$ACTARMCD)) %>%
-      rowwise() %>%
-      mutate(
-        STUDYID = pinfo$STUDYID,
-        SITEID = pinfo$SITEID,
-        USUBJID = pinfo$USUBJID,
-        CNSR = sample(c(0, 1), 1, prob = c(1-CNSR_P, CNSR_P)),
-        AVAL = rexp(1, LAMBDA),
-        AVALU = "DAYS",
-        EVNTDESC = if (CNSR == 1) sample(evntdescr_sel[-c(1:2)], 1) else sample(evntdescr_sel, 1)
-      ) %>% select(-LAMBDA, -CNSR_P)
+  ADTTE <- split(ADSL, ADSL$USUBJID) %>% # nolint
+    lapply(FUN = function(pinfo) {
 
-  })  %>% Reduce(rbind, .) %>%
+      lookup_TTE %>%
+        dplyr::filter(ARM == as.character(pinfo$ACTARMCD)) %>%
+        rowwise() %>%
+        mutate(
+          STUDYID = pinfo$STUDYID,
+          SITEID = pinfo$SITEID,
+          USUBJID = pinfo$USUBJID,
+          CNSR = sample(c(0, 1), 1, prob = c(1 - CNSR_P, CNSR_P)),
+          AVAL = rexp(1, LAMBDA),
+          AVALU = "DAYS",
+          EVNTDESC = if (CNSR == 1) sample(evntdescr_sel[-c(1:2)], 1) else sample(evntdescr_sel, 1)
+        ) %>%
+        select(-LAMBDA, -CNSR_P)
+
+    }) %>%
+    Reduce(rbind, .) %>%
     var_relabel(
       STUDYID = "Study Identifier",
       USUBJID = "Unique Subject Identifier"
     )
 
   # apply metadata
-  ADTTE <- apply_metadata(ADTTE, "metadata/ADTTE.yml", seed = seed, ADSL = ADSL)
+  ADTTE <- apply_metadata(ADTTE, "metadata/ADTTE.yml", seed = seed, ADSL = ADSL) # nolint
 
   ADTTE
-
 }