+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # random.cdisc.data 0.2.1
 
-* added cached versions of the datases `c<name>`
+* added cached versions of the datasets `c<name>`
 
 # random.cdisc.data 0.2.0